+++ src/server/routers/index.ts
@@ -3,10 +3,7 @@
  */
 import { publicProcedure, router } from '@/libs/trpc';
 
-import { configRouter } from './config';
-
 export const appRouter = router({
-  config: configRouter,
   healthcheck: publicProcedure.query(() => "i'm live!"),
 });
