+++ src/services/global.ts
@@ -1,10 +1,8 @@
-import { DeepPartial } from 'utility-types';
-
 import { dataSync } from '@/database/client/core';
-import { trpcClient } from '@/libs/trpc/client';
-import { LobeAgentConfig } from '@/types/agent';
 import { GlobalServerConfig } from '@/types/serverConfig';
 import { StartDataSyncParams } from '@/types/sync';
+
+import { API_ENDPOINTS } from './_url';
 
 const VERSION_URL = 'https://registry.npmmirror.com/@lobehub/chat';
 
@@ -20,11 +18,9 @@ class GlobalService {
   };
 
   getGlobalConfig = async (): Promise<GlobalServerConfig> => {
-    return trpcClient.config.getGlobalConfig.query();
-  };
+    const res = await fetch(API_ENDPOINTS.config);
 
-  getDefaultAgentConfig = async (): Promise<DeepPartial<LobeAgentConfig>> => {
-    return trpcClient.config.getDefaultAgentConfig.query();
+    return res.json();
   };
 
   enabledSync = async (params: StartDataSyncParams) => {
