+++ src/store/agent/slices/chat/initialState.ts
@@ -1,20 +1,14 @@
-import { DeepPartial } from 'utility-types';
-
 import { DEFAULT_AGENT_CONFIG } from '@/const/settings';
 import { LobeAgentConfig } from '@/types/agent';
 
 export interface AgentState {
   activeId: string;
-  agentConfig: DeepPartial<LobeAgentConfig>;
-  defaultAgentConfig: LobeAgentConfig;
+  agentConfig: LobeAgentConfig;
   isAgentConfigInit: boolean;
-  isDefaultAgentConfigInit: boolean;
 }
 
-export const initialAgentChatState: AgentState = {
+export const initialSessionState: AgentState = {
   activeId: 'inbox',
-  agentConfig: {},
-  defaultAgentConfig: DEFAULT_AGENT_CONFIG,
+  agentConfig: DEFAULT_AGENT_CONFIG,
   isAgentConfigInit: false,
-  isDefaultAgentConfigInit: false,
 };
